+++ neural_compressor/compression/pruner/__init__.py
@@ -19,3 +19,7 @@
 # model slim related
 from .model_slim.auto_slim import parse_auto_slim_config
 from .model_slim.auto_slim import model_slim
+
+
+
+
